+++ RAnalogs/RAnEn/R/readNc2.R
@@ -18,14 +18,23 @@
 #' 
 #' @param file The file to read
 #' @param recursive Whether to read variables recursively
+#' @param read_attrs Whether to read attributes. Make sure there are no
+#' name conflicts between attributes and variables. Only attributes under
+#' groups or in the root group are read. Attributes associated with variables
+#' are not read.
 #' 
 #' @md
 #' @export
-readNc2 <- function(file, recursive = F) {
+readNc2 <- function(file, recursive = F, read_attrs = F) {
 	
 	nc <- RNetCDF::open.nc(file)
 	results <- RNetCDF::read.nc(nc, recursive = recursive)
 	RNetCDF::close.nc(nc)
+	
+	if (read_attrs) {
+		attrs <- readNcAttrs(file, recursive = recursive)
+		results <- mergeList(results, attrs)
+	}
 	
 	return(results)
 }